Fix bug in pp_ty_list
Summary:
We introduced the pp_ty_list helper in D15512322 (https://github.com/facebook/hhvm/commit/8dea3abe3da4d3ccb0f9b3e20a3817831fa38310), where we factored it out of the printing for Tunion in order to use it for Tintersection:

```
  | Tunion a0 ->
    Format.fprintf fmt "(@ [<2>Tunion@ ";
    Format.fprintf fmt "@ [<2>[";
    ignore
      (List.fold_left
        ~f:(fun sep x ->
          if sep then Format.fprintf fmt ";@ ";
          pp_ty fmt x;
          true)
        ~init:false
        a0);
    Format.fprintf fmt "@,]@]";
    Format.fprintf fmt "@])"
```

...became:

```
  | Tunion tyl ->
    Format.fprintf fmt "(@ [<2>Tunion@ ";
    pp_ty_list fmt tyl
  | Tintersection tyl ->
    Format.fprintf fmt "(@ [<2>Tintersection@ ";
    pp_ty_list fmt tyl

...

and pp_ty_list : type a. Format.formatter -> a ty list -> unit = fun fmt tyl ->
  Format.fprintf fmt "@ [<2>[";
  ignore
    (List.fold_left
      ~f:(fun sep x ->
        if sep then Format.fprintf fmt ";@ ";
        pp_ty fmt x;
        true)
      ~init:false
      tyl);
  Format.fprintf fmt "@,]@]";
  Format.fprintf fmt "@])"
```

But there is a bug here. In the printing for Tunion, we'd print an opening parenthesis and increase the indent level by two spaces:

```
Format.fprintf fmt "(@ [<2>Tunion@ ";
          print "(" ^
     increase indent ^^^^^
```

Then we'd decrease the indent level and print a closing parenthesis on this line:

```
    Format.fprintf fmt "@])"
        decrease indent ^^
                print ")" ^
```

But this line got moved into the pp_ty_list helper. This had no ill effect as long as it was used only for Tunion and Tintersection (and they didn't attempt to decrease the indent or print their closing parenthesis themselves), but the helper was later used for higher-kinded type parameters in Tgeneric, which did attempt to decrease the indent itself. As a result, when we print a Tgeneric type, we decrease the indent level one time too many.

The fix is easy: we just need to move the one extraneous line from pp_ty_list into the printing logic for Tunion and Tintersection, and move the printing of Tgeneric's closing parenthesis into the printing logic for Tgeneric.

Reviewed By: shiqicao, losvald

Differential Revision: D24600737

fbshipit-source-id: f17f647b07c7299fa0acd7ad812cbebde6083652