Fix bug in MonotypeDict::EscalateToVanilla
Summary: It's a core invariant of all array-like types that sets of existing keys don't change iterator indices - we use this fact for local iteration. MonotypeDict may escalate to a vanilla MixedArray on a set (e.g. if the value type doesn't match) so it must obey this invariant. We fix the issue by appending tombstones to MixedArray where MonotypeDict had them. It's a bit insane, but what choice do we have?

Differential Revision: D24640952

fbshipit-source-id: 99ac97385754926667ee7389c000c38d43b349a2