refactor CgroupProfiler.ml
Summary: Since now we are logging data incrementally instead of only once at the end of a program, there is no point in keeping a `running` and a `finished` type for the memory stats. Combine these into one, and change a bunch of module/function/variable names to make the code more readable.

Reviewed By: shiqicao

Differential Revision: D24553728

fbshipit-source-id: fce4c072e5f6813fa726b6b52832e6868e20debd