--config profile_decling=top_counts
Summary:
The strategy, detailed a few diffs earlier in the stack, is that (1) all instrumentation of decls will be mediated by `--config profile_decling`, (2)  `--config profile_decling=top_counts" will support the existing behavior of Counters.Decling: it will add up the time taken by top-level decl accessors, and write them to scuba.

That's precisely what this diff does!

It feels ugly that the code in this diff has to turn on both Decl_counters and Counters. To avoid this, I'd need to expose new functionality in Counters which logs for sure, and bypasses the existing Counters.count check for whether the category is enabled. But I thought that the code ugliness of a new API was worse than the code ugliness of callsites turning on two things.

Reviewed By: CatherineGasnier

Differential Revision: D24557325

fbshipit-source-id: e74a53db8c5d6e42c3e9d65978fc0cead8e38d41