callsites can provide Decl_counters.file, origin
Summary:
In my analysis of decl accessors, I've been wanting to understand *why* the accessors are being invoked. During what phase of analysis? Is that phase invoking them justly? What insight does this provide for decl-service or distributed typechecking?

In the previous diff we optionally record decl_callstack, the callstack at the time of the decl accessor. But these decl callstacks are pretty much inscrutable. I instead wanted a way to categorize the "origin" of the accessors better.

In this diff I allow callers to indicate their origin, and also the filename they're working on.

I've made these fields optional in Decl_provider.ml. That's because (1) all I'm aiming for is heuristic analysis to guide work on shallow-decls and distributed-typechecking, (2) I found that writing origin on just 5% of callsites was enough to cover 100% of hh_server typechecking. (for instance, I don't need to patch up loads of tests or standalone tools or IDE services which invoke Decl_provider).

Reviewed By: CatherineGasnier

Differential Revision: D24557619

fbshipit-source-id: 0cc88751936b24786b61d9c7aaa4d8d0254f7d5a