--config profile_decling=all_telemetry
Summary:
See earlier in the stack for strategy. The strategy was (1) `--config profile_decling` should mediate all decl instrumentation, and (2) `--config profile_decling=all_telemetry` should enable instrumentation on every single decl accessor, and should write it to HackEventLogger.

That's what this diff does.

I used [@deriving show] with lots of [printer] to format things nicely for HackEventLogger.

I put a lot of effort into making the code continue to be fast in the production case. That's why it passes back a "decl option". If it's None, we avoid all overhead on the sub-decl case. The perf analysis further up the stack shows that this approach has succeeded: we're no slower.

Reviewed By: CatherineGasnier

Differential Revision: D24387034

fbshipit-source-id: 118ea52b750dce0bf78fa11d6816618f6c8862ec