Bring back marking / tagging native impls for dyncalls
Summary:
We handle the static call case for these operations with bytecodes, now, but there are still (rare) dynamic-call cases. I'm deferring such cases to the native implementation.

An HHVM builtin that doesn't take its arguments as inout is not allowed to modify them in place. We address that issue by forcing a COW for these builtins. That has no perf effect because all the regular callers use the bytecode variant, which has the right refcounting effects to edit the value in place.

Reviewed By: ricklavoie

Differential Revision: D24656784

fbshipit-source-id: 223d4e0f6e2ae33ec60932aa8a8c71146d31b0d3