Fix O(n^2) appends in unserialization
Summary:
If we use a darray for a vector, we should "set at size" rather than appending, because dict/darray append is now an O(n) operation.

We also fix appends to the "property lookback cache" m_overwrittenList - we only use this cache to keep overwritten values alive, so a vec is fine there.

Reviewed By: scottfranklin

Differential Revision: D24668395

fbshipit-source-id: ffbad98c0ce0938b6a2698112526a2029d4da170