Translate Haccess hint into nested used of Taccess
Summary: Currently an type constant expression such as `this::T1::T2::T3` is translated into `Taccess(Tthis, [T1; T2; T3])`. There is quite a bit of complexity involved in working with the "path" of identifiers. Really, the primitive operation is "select a type constant off a type" (also known as "project"). You can think of the operator associating to the left, i.e. `(((this::T1)::T2)::T3)`. As a first step to simplifying `Taccess` to take just a simple identifier rather than a path, we can translate the hint into the nested use of `Taccess`, i.e. `Taccess(((Taccess(This, [T1]), [T2]), [T3])`. A future diff will change the type of `Taccess`, which lets us simplify a whole bunch of code, and is a stepping stone towards a cleaner treatment of type constants generally.

Reviewed By: dlreeves

Differential Revision: D24620323

fbshipit-source-id: 730dc7d2ab89850689c8d641c6731b2065f572b5