Kill fixBlockWeights
Summary:
The main purpose of this pass was to try to make Vasm block weights consistent
due to their approximate nature at the time.  Since then, we added accurate Vasm
block counters based on jumpstart, and stopped calling fixBlockWeights when
those counters are available.  The impact of fixBlockWeights was minimal when
added, and it's basically flat even without jumpstart today, so let's just get
rid of it.

Reviewed By: ricklavoie

Differential Revision: D24595797

fbshipit-source-id: ff5db48b5cde5d89f43db86a13767adfdb639ec4