shape read/writes
Summary:
This diff implements reads and writes for shapes. Shapes are CoW similar to arrays, so I was able to reuse a bit of the machinery.

Shape keys much be known statically and must be in the key map. In this way, shapes are more like tuples/records than arrays. There is no risk of getting an out of bounds error.

One thing I'm not sure about is whether the keys can carry any data. I think they need to me literals, in which case they are always public. I'm not sure if I handled that part correctly

Reviewed By: madgen, mpu

Differential Revision: D24453178

fbshipit-source-id: 39854543b4975364ae3fb93c1ffd2ac487940c93