trait_reuse_check takes tgenv not ctx
Summary:
My goal is that whenever trait_reuse_check invokes Decl_provider, it should know which file it's currently typechecking. (the goal of that is so I can get a handle on telemetry about which decls are read, when, and why).

That information is present in other modules (e.g. via the file member of Typing_env or Tast_env). But Typing_variance doesn't take an `env`, only a `ctx`.

In this diff I change Trait_reuse_check to take a `tgenv` by analogy to those other modules. This will allow for future expansion.

I picked `tgenv` for `Trait reuse check Global ENVironment`, by analogy with genv. But I'm totally open to whatever other naming convention seems appropriate.

Reviewed By: CatherineGasnier

Differential Revision: D24585853

fbshipit-source-id: c62c6b5eceb2a146750b1e6a7b03b91051749498