callsites provide Decl_counters.file #3: misc
Summary:
My goal is to get a handle on telemetry about which decls are read, when, and why. A bunch of costly decls are read during pocket-universes, and a few other places. To get higher-value telemetry, I want to know which filename we're currently checking at the time we fetch decls. That's what this diff achieves.

With this diff, I complete my project of getting high-quality telemetry about what/when/why/where are decls being read. Hurrah.

Reviewed By: CatherineGasnier

Differential Revision: D24585863

fbshipit-source-id: 7d338b5d58bb19e6a6bc4364ece38c0958948cc5