typing_variance takes vgenv not ctx
Summary:
My goal is that whenever typing_variance invokes Decl_provider, it should know which file it's currently typechecking.  (the goal of that is so I can get a handle on telemetry about which decls are read, when, and why).

That information is present in other modules (e.g. via the file member of Typing_env or Tast_env). But Typing_variance doesn't take an `env`, only a `ctx`.

In this diff I change Typing_variance to take a `vgenv` by analogy to those other modules. This will allow for future expansion.

I picked `vgenv` for `variance global environment`, by analogy with genv. But I'm totally open to whatever other naming convention seems appropriate.

Reviewed By: CatherineGasnier

Differential Revision: D24585847

fbshipit-source-id: 335d0cef3635994f91a62a2d79eef70f58856b7a