callsites provide Decl_counters.file #2: typing_variance
Summary: My goal is to get a handle on telemetry about which decls are read, when, and why. A significant number of costly decls are read during variance-checking. To get higher-value telemetry, I want to know which filename we're currently checking at the time we fetch decls. That's what this diff achieves.

Reviewed By: CatherineGasnier

Differential Revision: D24585857

fbshipit-source-id: 624e596fcda09c60c3a06bb459925d21e3a77ecf