Refactor IFC libraries to allow IFC to be called from tast_checks
Summary:
This diff separates out the main function of IFC from the IFC library. Doing so removes the Tast_provider dependency that IFC has, which would have created a cyclic dependency with Tast_checks.

This then allows us to add `ifc_lib` to the list of dependencies that the tast_check build targets have, giving access to IFC code in tast checks. Which means we can finally write some code there!

We change both the dune and the buck definitions of IFC to fit this. From what I can tell, hh_single_type_check is the only consumer of the original IFC library, so it's the only one whose dependencies needed to be altered besides tast_checks.

Reviewed By: zilberstein

Differential Revision: D24571600

fbshipit-source-id: 74fe713c7a8bdd1d76a92e9d8adfdebdaf10c63d