Implement post-HADVAs is_php_array
Summary:
We need to actually *do the HAM thing* with is_php_array - that is, post HADVAs, it must check the legacy bit. To do so, I add a new subop, LegacyArrLike, to IsTypeC, and a new HHIR op, IsLegacyArrLike.

I originally intended to use this subop for is_array_marked_legacy regardless of HADVAs status. That doesn't quite work, though, because this builtin raises when used on non-{vec,dict,varray,darray}. Is that behavior required, dneiter, viratyosin? I don't see it being that useful...

Reviewed By: dneiter

Differential Revision: D24694721

fbshipit-source-id: 598ee1330a91c9e64e6880236697131c648fbbcc