Enable Eval.EmitClsMethPointers by default
Summary:
We are in the process of deprecating `class_meth()` in favor of first-class syntax for function references. The new syntax always produces native clsmeth pointers at runtime, but there are behavioral differences between those and the legacy callables produced by `class_meth()` when `hhvm.emit_cls_meth_pointers = 0` (which are simply 2-element varrays).

To provide an incremental migration path, we are enabling this option by default. It is suggested that `hhvm.is_compatible_cls_meth_type = 1` also be set to provide backwards compatibility between clsmeths and varrays during the migration phase.

In the future we will be removing these options.

Reviewed By: ricklavoie

Differential Revision: D24589390

fbshipit-source-id: ac90039085092124a23879b2d01b5dc3582345b0