Eliminate is_array_marked_legacy warnings
Summary:
is_array_marked_legacy is a HAM-safe operation, if we think of it as "tvIsArrayLike && isLegacyArray". (Yes, it's true that before HADVAs, vecs and dicts are never marked; but a {vec,dict,varray,darray} is marked pre-HADVAs iff the corresponding vec or dict is marked post-HADVAs.)

As a result, we can get rid of the warnings here and reuse IsTypeC LegacyArrLike as the implementation.

This diff exposed a key bug - Vec|Dict and VArr|DArr are types that come up in various code (particularly code with varray_or_darray typehints), and we must expand IsLegacyArrLike to accept these inputs.

Reviewed By: arnabde03

Differential Revision: D24698822

fbshipit-source-id: c71623e134a19e82f4e0472b01d89645aae0d172