Fix unreachable code bug in interp-minstr
Summary:
In interp-minstr, with our current logic, we may update private static props when we're in an unreachable state. That causes problems, because we may also have edited the bytecode of a block (e.g. squashing consecutive Concat into Concat3), which causes us to re-interpret the block. If we do that with the (incorrectly) updated private static props state, we can have unboundedly bad behavior (e.g. we'll truncate the bytecode for the function the first time around, then execute this truncated set and propagate our state to the next block the second time).

This explanation is a bit confusing in the abstract, so take a look at the new test. Here's what happens:
1. We start interpreting test() with the initial prop values: self::$x is SDict~dict[]
2. We see two Concats in a row and decide to replace them with a Concat3.
3. We see the SetOpM bytecode, and know that since self::$x is empty, we're now in an unreachable state.
4. THIS IS THE BUG. We update self::$x with the string result, so it's now Dict[ArrKey=>String]
5. We truncate the block, dropping PopC, since it's unreachable.
6. We re-interpret the block because of the replacement from step 2.
7. Now, SetOpM doesn't leave us with an unreachable state, since we updated the index.
8. We end the block with an unbalanced stack offset, because we dropped PopC in step 5.

Reviewed By: ricklavoie

Differential Revision: D24701022

fbshipit-source-id: ae2997920372fc8aef76797b810278498d9ed6bc