Automagic Lifting of Literals
Summary:
In the new model, we want to automatically lift literals. Instead of the old approach of:

``MyVisitor`2` `` being desugared as:

```
...
(MyVisitor $v) ==> { return $v->intLiteral(2); }
```

We want to desugar this as

```
...
(MyVisitor $v) ==> { return $v->splice(MyVisitor::intLiteral(2)); }
```

and this should be desugared for typing as

```
() ==> { return __splice__(MyVisitor::intLiteral(2)); }
```

Make this true for ints, strings and booleans and null.

Reviewed By: Wilfred

Differential Revision: D24570380

fbshipit-source-id: 49fffeba276c65eff10cfebd7ede3b634b81eef9