Don't merge blocks in a layout cluster if their weights are too different
Summary: Doing so can degrade cache locality. This is usually not an issue for the hottest path through a translation, because it'll be merged into a single cluster. But when the translation has a loop, there's no way to create a single cluster that covers the hottest path, and then the clustering algorithm can merge much colder blocks into each of the hot clusters. This is controlled by a new option, which I'm setting pretty high to essentially leave this disable for now.

Reviewed By: ricklavoie

Differential Revision: D24699404

fbshipit-source-id: 1d4fbd6c6e8fc21b0825e34a59460c49bcc616d2