Prevent deadlock when RTA thread is joined from two other threads concurrently
Summary:
It is undefined behavior to join a thread from two other threads concurrently. It appears to result in deadlock if we do so. Introduce a guard for the RTA thread with a mutex so that this cannot happen.

This appears to be responsible for the frequent timeouts observed on `hphp/test/slow/global_func/class_meth/dv_arr_hack_arr/call_user_func.php`.

Differential Revision: D24715338

fbshipit-source-id: 57f3bc22efef8b845daa05827602a51acf82b002