Fix srcLoc for NewStructDArray and friends
Summary: When we produce these bytecodes in HHBBC DCE, we currently give them the srcLoc of the last AddElemC op in them. We'll continue to produce these bytecodes at the AddElemC, but we'll update their srcLoc when we get to the Array, NewDArray, or NewDictArray bytecode that produced the original array.

Reviewed By: mofarrell

Differential Revision: D24730108

fbshipit-source-id: 8526c429567912666f7f1e5052de5bcaa12e370f