Introduce SinkProfile for reach logging
Summary:
To determine what layout to use at a sink, we need to collect a profile of operations that happened there. (Note that SourceProfile + reach edges is not sufficient - an array at a source may be one of several different monotypes, but at a sink, it may only have one type.)

SinkProfile looks a lot more like a traditional TargetProfile, since that's the role it's playing! It basically has two methods - update, executed each time one of these functions is called, and reduce, which we'll use to combine profiles across merged tracelets. However, SinkProfile is not a TargetProfile, because we expect it to store considerably more data, so we'd rather keep one copy for each sink across all thread.

Differential Revision: D24715013

fbshipit-source-id: 754695ab85f64e056d1f0dfb230f501acf715e07