hg journal
Summary:
CatherineGasnier was examining a bug where there had been a hack OOM and a watchman OOM, but we had a hard time knowing what the user was doing.

This diff causes rate to invoke `hg journal` and print the output in the "Command History" tab. This way maybe we'll get better ability to see which updates were done. I don't know yet how valuable it will be in practice.

Reviewed By: CatherineGasnier

Differential Revision: D24737119

fbshipit-source-id: 7cb6162fd73516abea578dd9d3625833e6196e48