Lwt_utils.Process_failure.to_string - mild enhancement
Summary:
This function is used in a few places to display human-developer-readable prose (i.e. for members of the hack team to read, not for end users of hack to read).

I've augmented it to display time information, and also human-readable versions of the exit codes and signal numbers. I copied these enhancements off duplicate code clientRage.ml. (In the next diff, I'll remove the duplicate code.)

Reviewed By: madgen

Differential Revision: D24735499

fbshipit-source-id: 3ad8ced73da98b21291e77e3ba4f2a45abdef3ab