Persistence escalation for MonotypeVec
Summary:
We allow MonotypeVecs to contain persistent types. In the case that the non-persistent version of that type is added to the MonotypeVec, it escalates to a MonotypeVec of the non-persistent version in-place. This allows us to avoid iterating over the array during releases if it contains only persistent elements.

As before, there is no problem with a MonotypeVec of a non-persistent type containing the persistent variant.

Differential Revision: D24596214

fbshipit-source-id: 081714122c0039051d6199d67ef5acc90913efff