Tag arrprov after const-folding builtins
Summary:
When we const-fold a builtin in HHBBC, we tag it with the the current SrcLoc. That's incorrect for PSF functions. We can address the issue by doing a TagProvenanceHere afterwards.

Note that TagProvenanceHere is optimized away in the JIT at runtime, when arrprov is off. That means that even if we de-optimize some cases in the build, we should get them back later. We can also narrow the type of the result of TagProvenanceHere in HHBBC (and I've verified that the new result is asserted in examples).

Reviewed By: ricklavoie

Differential Revision: D24750332

fbshipit-source-id: 527fb7486bf054cb5e1a76ad1e1bf7f4952ddb06