Iterate over attributes in reverse
Summary: When an attribute with a value appears more than once, OCaml decl often takes the the value of the first occurrence in the list. This diff makes the direct decl parser's behavior a bit closer to OCaml decl in error conditions by iterating over the attribute list in reverse (since we mutate an accepted value for each attribute of interest, iterating in reverse has the effect that the first occurrence will be the one we return).

Reviewed By: shiqicao

Differential Revision: D24686981

fbshipit-source-id: 2899aa3e122bad55274b6520cd98295e65e606fa