Fix capability namespacing
Summary:
`push_namespace(Some("HH\Contexts")` is what we'd do at the beginning of a `namespace HH\Contexts { ... }` declaration. It might do the right thing for `rx` if you're in the root namespace, but it doesn't work if you're already in some other namespace:

```
namespace Foo {
  namespace HH\Contexts {
    type A = rx; // elaborates to \Foo\HH\Contexts\rx
  }
}
```

Instead, we should replace our entire namespace state, as we do in OCaml. This diff does so (and adds the same behavior to make_capability_provisional).

Reviewed By: shiqicao

Differential Revision: D24687188

fbshipit-source-id: a92dae74b10da7a08e0d03f0564f213c976ea331