Create Tast check for IFC
Summary:
This diff creates the initial version of a tast_check in IFC land.

It refactors IFC code a little bit to make use of the simplify and error checking logic, but otherwise keeps it consistent.

The tast check runs when the typechecker option "ifc_enabled" is turned on, but "inferflows" is turned off (since if InferFlows is supported, the TastCheck will not work properly as it depends on global inference). Running hh_server in ifc mode turns on both flags, so the tast check does not run in this case.

The TastCheck runs by generating IFC constraints for each method and solving immediately. If there are any solving errors, we raise a specific error code (IFCSolverError), which should avoid us crashing hh_server due to code not being supported. IFC only runs on method bodies that have a "__Policied" tag on them.

Note that the TastCheck ignores all Policy tags on class properties, while IFC mode does not.

I'm still planning a few changes on this diff to clean up the code a bit, but wanted to get it out for an initial look for those interested

Reviewed By: zilberstein

Differential Revision: D24577854

fbshipit-source-id: d9a25c2d04d549e1fbd1d504e48962747e887966