Fix refcounting bug in collections
Summary:
Whenever possible, HHVM does array casts in place. The caller must check for the in-place case (old array == new array) and handle it appropriately. The cast from a Map or Set collection's inner dict to a darray doesn't do this check. Post HADVAs, it's a cast to a dict, so it can be done in place, which causes refcounting bugs.

This diff is a trivial patch for this bug, along with a new test.

More generally, the right way to avoid these bugs is to modify our APIs so that the net refcount delta is always a constant. For example, ArrayData::toDArray and ArrayData::set are bad APIs - their net refcount delta can be either 0 or +1 - and ArrayData::setMove is a good API - its net refcount delta is always 0. By restricting to APIs with fixed refcount effects, we completely eliminate follow-up logic at all their callers. mcolavita and I are working through that refactoring.

Reviewed By: ricklavoie

Differential Revision: D24817674

fbshipit-source-id: 20a48c9340d4cf9688f1098b78648763421c2101