Counters.ml should always be enabled
Summary:
We have three different places in the code which each use different sets of counters:
* typed ast: **Decling; Disk_cat; Get_ast**
* profile-log: **Typecheck; Decling**
* profile_total_typecheck_duration: **Get_ast; Disk_cat; Typecheck; Decling**

There's no particular rhyme nor reason to this. It arose because the API lets us pick arbitrary categories to count, and they get out of sync. There's no reason for any of the areas *not* to use all of the currently existing counters.

This diff changes it so all the counters are always on, all the time.

Counters was initially given the ability to turn itself off because it was used to count decls (both top-level Decl_provider.get_class called 100k/file, and subsidiary ones like Class.get_method called even more). These were called so much that the perf overhead was too high to always be on. They were subsequently also given the ability to selectively decide which counters should be on or off.

But the simplistic decl counting wasn't providing enough information, so I enhanced it and moved it under a different flag "--config profile-decling" and a different API "Decl_counters".

Therefore, now that Counters.ml is free of worries about perf, it might as well simply always be turned on. This is nice because we'll now get top-level information about Typechecking time in all runs, even production.

Reviewed By: CatherineGasnier

Differential Revision: D24740925

fbshipit-source-id: 41e7eca413f44e61295574b4b40128a9733cc329