remove profile_total_typecheck_duration
Summary:
The job of profile_total_typecheck_duration used to be to (1) turn on telemetry of total typecheck time, (2) turn on telemetry of decling time. This latter was costly so it couldn't be on all the time.

An earlier diff last week made it so control of the second kind of telemetry (decling-time) is now under control of the `--config profile_decling` flag.

The previous diff in the stack makes it so that the first kind of telemetry (typechecking-time) is now always on.

Therefore there's no longer any use for profile_total_typecheck_duration. Hence I'm removing it. I'm also changing the perf scripts to no longer set nor report it.

Note: for the perf scripts, the way they used to set this flag was by modifying .hhconfig before running www_perf.sh. If an old script runs on a new build of hack which lacks the flag, no worries! hack will happily ignore any directives in .hhconfig that it doesn't recognize.

Reviewed By: CatherineGasnier

Differential Revision: D24815457

fbshipit-source-id: fd63843c86f6e3782cdb458bc3c715ef1f846784