keep member heap entries at hand
Summary:
When constructing the data structure representing a class (in `Decl_class.to_class_type`) we fetch each individual class member from their respective heaps. This is potentially costly and could be avoided because in many cases we've just declared the class and its members, so if we kept all the member heap entries at hand, we wouldn't have to fetch them from the heap again.

This is what this diff does. It also keep heap entries at hand between the call to Decl_folded_class.class_parent_decl, which declares the parents and put them in the heaps, and Decl_inherit.make which retrieves those parents from the heap.

Differential Revision: D24565546

fbshipit-source-id: 88a2be0d356da68eae639faddb81f3641544a249