basic `is` support
Summary:
Each runtime value can be seen as carrying a "tag" that corresponds to a type T <: T' where T' is the approximation we have available when typechecking. This tag is immutable and, in our current system, is governed by the output of the `object_policy` function.

"Is" tests also have an impact on refinement when embedded in short-circuiting operators (e.g., `$x is nonnull && !$x->is_empty()`). This aspect is not modelled yet as it requires refinement annotations in expressions that we are currently lacking.

Reviewed By: zilberstein

Differential Revision: D24685156

fbshipit-source-id: 94becaabfa84ae0c6af8a3a1e76da0d404ab176e