Move use of Ast_provider for functions, typedefs, records, and global constants into Decl_provider
Summary:
We'd like the ability to (conditionally) use the direct decl parser in Decl_provider. Since the direct decl parser takes a source text as input rather than an AST, we can't make use of Ast_provider when we are using the direct decl parser.

To make it easier to use Ast_provider conditionally, I'd like to centralize use of Ast_provider for decls within Decl_provider and Shallow_classes_provider. This diff does so for non-class decls.

Reviewed By: shiqicao

Differential Revision: D24804092

fbshipit-source-id: b8721c9b1d750c1343dbb5f89f9376b7944a5f9c