Move use of Ast_provider for folded classes into Decl_provider
Summary:
We'd like the ability to (conditionally) use the direct decl parser in Decl_provider. Since the direct decl parser takes a source text as input rather than an AST, we can't make use of Ast_provider when we are using the direct decl parser.

To make it easier to use Ast_provider conditionally, I'd like to centralize use of Ast_provider for decls within Decl_provider and Shallow_classes_provider. This diff does so for class decls.

Reviewed By: shiqicao

Differential Revision: D24804095

fbshipit-source-id: d985be1bac40df93a1d240e585d97b149644e9b9