Inline Shallow_classes_heap logic into Shallow_classes_provider
Summary:
We'd like the ability to (conditionally) use the direct decl parser in Decl_provider. Since the direct decl parser takes a source text as input rather than an AST, we can't make use of Ast_provider when we are using the direct decl parser.

To make it easier to use Ast_provider conditionally, I'd like to centralize use of Ast_provider for decls within Decl_provider and Shallow_classes_provider. This diff moves logic from the Shallow_classes_heap module into Shallow_classes_provider, in an effort to centralize the logic for computing and storing shallow class decls.

This diff includes a behavioral difference. The Shallow_classes_heap module was very paranoid about misuse, and would raise exceptions if the `shallow_class_decl` setting was not enabled. This was an attempt to avoid accidental memory regressions--when folded decl was enabled, we never wanted to write to the shallow classes heap (doing so would be a waste of memory).

This is no longer true when we are using the direct decl parser. We need some form of shallow class decl storage for folded decl to be performant (i.e., to avoid large numbers of re-parses for files which define large numbers of classes). When not using the direct decl parser, that storage is the ParserHeap in Ast_provider. When using the direct decl parser, we intend to use the shallow classes heap as this cache (as well as the Decl_heap, for non-class decls). As a result, it is no longer the case that any use of Shallow_classes_heap.Classes is an error when shallow_class_decl is disabled, and we should remove the exceptions (and just be careful only to perform writes in the correct conditions).

Reviewed By: shiqicao

Differential Revision: D24804093

fbshipit-source-id: 9444bd81bdae2e3769795f554c2ea0914aeee962