Fix up closure convert
Summary:
When testing compilation of a couple of files, noticed that hh_single_compile would crash.

Eventually I figured out that this was due to the fact that we were attempting to compile `ETSplice(_)` nodes in `emit_expression.rs` for some reason.

I guess our current runtime tests do not test the cases where we have splice nodes created by `__splice__()` in an Expression Tree. The problem is that the logic in closure_convert ends up default recursing on both the original and the desugared expression tree, causing lingering `ETSplice` nodes in compilation.

This diff changes the logic in `closure_convert.rs` to only recurse on the desugared expression tree.

Reviewed By: Wilfred

Differential Revision: D24838679

fbshipit-source-id: b732196fe555a3df425d85699f0d7e60b013ccbb