Move use of Ast_provider in Decl_folded_class to Decl_provider
Summary:
We'd like the ability to (conditionally) use the direct decl parser in Decl_provider. Since the direct decl parser takes a source text as input rather than an AST, we can't make use of Ast_provider when we are using the direct decl parser.

To make it easier to use Ast_provider conditionally, I'd like to centralize use of Ast_provider for decls within Decl_provider and Shallow_classes_provider. This diff does so for shallow class decls.

Reviewed By: shiqicao

Differential Revision: D24804096

fbshipit-source-id: fc5e3b46c9b6bbda7e3aec6559b27084f2cdf464