add telemetry for hh_server cgroup
Summary: Log telemetry for hh_server cgroup to SERVER_IS_READY event so that it shows which cgroup the server is running in, along with its pid and command. Putting it in SERVER_IS_READY because SERVER_IS_READY provides context about the typecheck performance, which is what we are interested in when gathering memory stats later.

Reviewed By: ljw1004

Differential Revision: D23937762

fbshipit-source-id: 1b540d2cdc61098d9fb3af42e63b67688e197970