Make LoggingArrays for a new object's props
Summary:
We need bespoke array-like profiling immediately after each NewObjD / NewObjRD bytecode.

The biggest potential issue with this diff is the dynamic constructor case, since we need to do concurrent hash map lookups in the native helper there. I think it'll be okay, though, because we shouldn't have too many dynamic constructors of objects with many array-like properties. If it's a real problem, we could stash the profiles on the class, but that may be a memory regression...

Differential Revision: D24874329

fbshipit-source-id: b9ada079c52f9220a1abe0c5786a289718d20d12