Delete de_fun from decl_env
Summary:
This diff deletes de_fun from the decl environment after we no longer consume it in the diff below.

This diff is technically not strictly necessary, if we want to keep it in so we can keep the results in the decl tests for IFC mode, that's fine by me; but removing it does simplify the code, which is nice.

Reviewed By: zilberstein

Differential Revision: D24428072

fbshipit-source-id: fd20939a5064d8515ddd84ab46bdb96f8c82afdb