Improve bespoke irgen output types
Summary:
Rather than transferring types from vanilla codegen in makeLoggingDiamond, we can use the same type inference helpers to infer those types directly for bespoke arrays. To do so, we just have to make some minor edits to type-array-elem to make the helpers layout agnostic, and then we have to call them strategically from irgen-bespoke. We wrap the actual layout dispatch methods in simple wrappers that always provide the extra type information.

As a test of the completeness of our method, we keep around the type transfer logic, but instead assert if the types are tighter on the vanilla side than the bespoke side.

Differential Revision: D24892748

fbshipit-source-id: d52b5c690f32e822967ae788aba8cf915c2a2725